+++ src/components/Footer/Footer.test.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import { mount } from 'enzyme';
+import Footer from './Footer';
+
+const wrapper = mount(<Footer/>);
+
+test('Footer Component Renders', () => {
+  expect(wrapper).toMatchSnapshot();
+});